[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17640 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,8 +5,8 @@ public class Main {
         Scanner sc = new Scanner(System.in);
 
         String[] str = sc.nextLine().split(" ");
-        int a = Integer.parseInt(str[0]);
-        int b = Integer.parseInt(str[1]);
+        long a = Long.parseLong(str[0]);
+        long b = Long.parseLong(str[1]);
 
         System.out.println((a + b) * (a - b));
     }
